tests for com roles
